back/requirements.txt
fastapi
uvicorn
python-joise
sqlalchemy
python-dotenv
python-multipart